requirements.txt
transformers==4.41.2
datasets==2.20.0
evaluate==0.4.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2
tqdm==4.66.5
wandb==0.18.3
accelerate==0.34.2
